+++ tests/functional/test_health.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 
 import json
 import time
+from typing import Any
 
 import pytest
 import tango
@@ -155,25 +156,79 @@ def device_online(
     change_event_callbacks.assert_change_event("device_state", Anything)
 
 
-@given("the Station has been commanded to turn on")
-@when("the Station has been commanded to turn on")
-def station_on(device_proxies: dict[str, tango.DeviceProxy]) -> None:
+@pytest.fixture(name="command_info")
+def command_info_fixture() -> dict[str, Any]:
     """
-    Command the station to turn on.
+    Fixture to store command ID.
+
+    :returns: Empty dictionary.
+    """
+    return {}
+
+
+@given(parsers.cfparse("the {device} has been commanded to turn On"))
+@when(parsers.cfparse("the {device} has been commanded to turn On"))
+def device_on(
+    device: str,
+    device_proxies: dict[str, tango.DeviceProxy],
+    command_info: dict[str, Any],
+) -> None:
+    """
+    Command the device to turn on.
+
+    :param device: device to turn on.
+    :param device_proxies: dictionary of device proxies.
+    :param command_info: dictionary to store command ID.
+    """
+    _, [command_id] = device_proxies[device].On()
+    command_info[device + "On"] = command_id
+
+
+@given(parsers.cfparse("the {device} has been commanded to turn to Standby"))
+@when(parsers.cfparse("the {device} has been commanded to turn to Standby"))
+def device_standby(
+    device: str,
+    device_proxies: dict[str, tango.DeviceProxy],
+    command_info: dict[str, Any],
+) -> None:
+    """
+    Command the device to turn on.
+
+    :param device: device to turn on.
+    :param device_proxies: dictionary of device proxies.
+    :param command_info: dictionary to store command ID.
+    """
+    _, [command_id] = device_proxies[device].Standby()
+    command_info[device + "Standby"] = command_id
+
+
+@then(parsers.cfparse("the {device} {command} command finishes"))
+def device_command_finishes(
+    device_proxies: dict[str, tango.DeviceProxy],
+    device: str,
+    command: str,
+    command_info: dict[str, Any],
+) -> None:
+    """
+    Wait for a command to complete.
 
     :param device_proxies: dictionary of device proxies.
+    :param device: device under test.
+    :param command: command to wait for
+    :param command_info: dictionary to store command ID.
     """
-    device_proxies["Station"].On()
-
-
-@given("the Station has been commanded to turn to standby")
-def station_standby(device_proxies: dict[str, tango.DeviceProxy]) -> None:
-    """
-    Command the station to turn to standby.
-
-    :param device_proxies: dictionary of device proxies.
-    """
-    device_proxies["Station"].Standby()
+    command_id = command_info[device + command]
+    count = 0
+    while (
+        device_proxies[device].CheckLongRunningCommandStatus(command_id) != "COMPLETED"
+    ):
+        time.sleep(1)
+        count += 1
+        if count > 10:
+            pytest.fail(
+                f"{device}.{command} did not complete: "
+                f"{device_proxies[device].CheckLongRunningCommandStatus(command_id)}"
+            )
 
 
 @given(parsers.cfparse("the {device} reports that its {attribute} is {value}"))
@@ -255,3 +310,26 @@ def set_subrack_health_params(device_proxies: dict[str, tango.DeviceProxy]) -> N
     }
     subrack_device = device_proxies["Subrack"]
     subrack_device.healthModelParams = json.dumps(new_board_params)
+
+
+@then(parsers.cfparse("the {device} reports that it is {programming_state}"))
+def check_device_programming_state(
+    device_proxies: dict[str, tango.DeviceProxy], device: str, programming_state: str
+) -> None:
+    """
+    Check the tileProgrammingState of a device.
+
+    :param device_proxies: dictionary of device proxies.
+    :param device: The device to check. Must be Station or Tile.
+    :param programming_state: The programming state to check for.
+    """
+    # List from Station, str from Tile.
+    device_programming_state = device_proxies[device].tileProgrammingState
+    if isinstance(device_programming_state, list):
+        # Assert all Tiles in the Station are in the specified state.
+        assert all(
+            tile_programming_state == programming_state
+            for tile_programming_state in device_programming_state
+        )
+    elif isinstance(device_programming_state, str):
+        assert device_programming_state == programming_state